A number of changes/improvements - see notes
Reworked how hashed files are tracked - previously, a single string variable kept track of the last hashed file in order to add it to the data grid. Now each hashed file is added to a queue which is processed in "batches," for lack of a better term. This allows for better threading support since a single hash variable string coule change before being added to the datagrid

Added an option to ignore files with a size of zero bytes

Added an option to enable multithreaded processing of file comparisons

File sizes now appear in KB instead of MB

-Issues
Occasional crashes when using multitheaded file comparisons. IndexOutOfRangeException. Issue possibly lies in the fact that the internal list for the ConcurrentDictionary in FileDictionary is not concurrent itself
Duplicate file counts are often wildly inaccurate, on the low side